Basic Set Up.
Added a main.tscn and world.tscn. main.tscn serves the function of loading world.tscn. It also holds nodes, who are meant to be accessed from anywhere. In this case a QuitOnInput and a FullscreenOnInput node.
Using Quarternius' simple nature pack, a simple collection of nature is added to world.tscn, as well as ground.
Using PlayerController as a base, a player is added to world.tscn as well.