Music and World Generation.
Music now starts playing after 5 seconds  when entering the game.
World generation is based on 16x16x16 chunks and is meant to be generated in three dimentions. Rn the generation generates all chunks within a render distance at _ready(). Then, when the player enters a new chunk, it renders all the chunks at the edge of the render distance towards which the player is moving. (For example betwen (-2, -2, 5) and (2, 2, 5) if the player is walking in a +z direction.)